@@ -10,24 +10,25 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@paypal/react-paypal-js": "^8.5.0",
-    "axios": "^1.6.5",
-    "papaparse": "^5.4.1",
-    "react": "^18.2.0",
-    "react-datepicker": "^7.5.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.21.2",
-    "react-select": "^5.8.0",
+    "@paypal/react-paypal-js": "^8.9.2",
+    "axios": "^1.12.2",
+    "dotenv": "^17.2.3",
+    "papaparse": "^5.5.3",
+    "react": "^19.2.0",
+    "react-datepicker": "^8.7.0",
+    "react-dom": "^19.2.0",
+    "react-router-dom": "^7.9.3",
+    "react-select": "^5.10.2",
     "tradestracker": "file:.."
   },
   "devDependencies": {
-    "@types/react": "^18.2.43",
-    "@types/react-dom": "^18.2.17",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "eslint": "^8.55.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
-    "vite": "^5.0.8"
+    "@types/react": "^19.2.0",
+    "@types/react-dom": "^19.2.0",
+    "@vitejs/plugin-react-swc": "^4.1.0",
+    "eslint": "^9.36.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^6.1.0",
+    "eslint-plugin-react-refresh": "^0.4.23",
+    "vite": "^7.1.9"
   }
 }
