@@ -1,13 +1,13 @@
-const Import = require('../models/TradesImportModel.js')
-const asyncHandler = require("../middleware/asyncHandler.js")
+import Import from '../models/TradesImportModel.js'
+import asyncHandler from "../middleware/asyncHandler.js"
 
-const saveData = asyncHandler(async (req, res) => {
+export const saveData = asyncHandler(async (req, res) => {
   const { ticker, candles } = req.body
   const savedData = Stock.create({ ticker, candles })
   res.status(200).json(savedData)
 })
 
-const getData = asyncHandler(async (req, res) => {
+export const getData = asyncHandler(async (req, res) => {
   const { ticker } = req.params
   const foundData = await Stock.findOne({ ticker: ticker })
   if (foundData) {
@@ -18,5 +18,3 @@ const getData = asyncHandler(async (req, res) => {
     throw new Error('Stock data not found')
   }
 })
-//
-module.exports = { saveData, getData }
