@@ -1,19 +1,25 @@
-require('dotenv').config()
-const path = require('path')
-const express = require("express")
+
+import 'dotenv/config'
+import express from "express"
+import path from 'path'
 const port = process.env.PORT || 3000;
-const cors = require('cors')
-const mongoose = require('mongoose')
-const cookieParser = require('cookie-parser')
+import cors from 'cors'
+import mongoose from 'mongoose'
+import cookieParser from 'cookie-parser'
 
-const errorHandler = require('./middleware/errorHandler.js')
-const tradesRoutes = require("./routes/tradesRoutes.js")
+import errorHandler from './middleware/errorHandler.js'
 
 
-const accountsRoutes = require('./routes/accountsRoutes.js')
-const profileRoutes = require('./routes/profileRoutes.js')
-const stockDataRoutes = require('./routes/stockDataRoutes.js')
-const authRoutes = require('./routes/authRoutes.js')
+
+import accountsRoutes from './routes/accountsRoutes.js'
+import profileRoutes from './routes/profileRoutes.js'
+import stockDataRoutes from './routes/stockDataRoutes.js'
+import authRoutes from './routes/authRoutes.js'
+import paymentRoutes from './routes/paymentRoutes.js'
+import tradeRoutes from './routes/tradeRoutes.js'
+import tradeAccountRoutes from './routes/tradeAccountRoutes.js'
+
+
 
 const app = express();
 app.use(express.json({ limit: '50mb' }))
@@ -29,10 +35,14 @@ const logRequestStart = (req, res, next) => {
 app.use(logRequestStart)
 
 
+
 app.use('/api/accounts', accountsRoutes)
 app.use('/api/profile', profileRoutes)
 app.use('/api/auth', authRoutes)
 app.use('/api/stocks', stockDataRoutes)
+app.use('/api/paypal', paymentRoutes)
+app.use('/api/trade', tradeRoutes)
+app.use('/api/trade-account', tradeAccountRoutes)
 
 
 const __dirname2 = path.resolve()
@@ -42,7 +52,7 @@ const __dirname2 = path.resolve()
 app.use(express.static(path.join(__dirname2, '/frontend/dist')))
 
 // redirect non /api routes to index.html
-app.get('*', (req, res) => {
+app.get('/*path', (req, res) => {
   res.sendFile(path.resolve(__dirname2, 'frontend', 'dist', 'index.html'))
 })
 
