@@ -1,12 +1,14 @@
+import { Link } from "react-router-dom";
+
 const Footer = () => {
   return (
     <div className="footer-section">
       <div className="container">
         <div className="top-footer">
           <div className="left-footer">
-            <img className="footer-logo" src="../images/quantsky.svg" alt="" />
+            <img className="footer-logo" src="../images/darkso.svg" alt="" />
             <div className="social-icons-section">
-              <img
+              {/* <img
                 className="social-svg"
                 src="../images/social-facebook.svg"
                 alt=""
@@ -21,12 +23,19 @@ const Footer = () => {
                 src="../images/social-instagram.svg"
                 alt=""
               />
-              <img className="social-svg" src="../images/social-x.svg" alt="" />
-              <img
-                className="social-svg"
-                src="../images/social-youtube.svg"
-                alt=""
-              />
+              <img className="social-svg" src="../images/social-x.svg" alt="" /> */}
+              <Link
+                to={"https://www.youtube.com/@karlos.trades"}
+                target="_blank"
+                className="youtube-btn"
+              >
+                <img
+                  className="social-svg"
+                  src="../images/social-youtube.svg"
+                  alt=""
+                />
+                YouTube
+              </Link>
             </div>
           </div>
           <div className="right-footer">
@@ -47,16 +56,17 @@ const Footer = () => {
           </div>
         </div>
         <div className="bottom-footer">
-          <span>© 2025 QuantSky LLC. All rights reserved.</span>
+          <span>© 2025 SignalOverwatch.com LLC. All rights reserved.</span>
           <span>
-            QuantSky LLC (“QuantSky”) is a software as a service provider and
-            financial education company. QuantSky does not provide investment
-            advice, guidance or recommendation, and QuantSky does not provide
-            commodity trading advice based on, or tailored to, the commodity
-            interest or cash market positions or other circumstances or
-            characteristics of particular clients, nor does QuantSky direct any
-            client accounts. No information set forth on this website is an
-            invitation to trade any specific investments. QuantSky does not take
+            SignalOverwatch.com LLC (“SignalOverwatch.com”) is a software as a
+            service provider and financial education company.
+            SignalOverwatch.com does not provide investment advice, guidance or
+            recommendation, and SignalOverwatch.com does not provide commodity
+            trading advice based on, or tailored to, the commodity interest or
+            cash market positions or other circumstances or characteristics of
+            particular clients, nor does SignalOverwatch.com direct any client
+            accounts. No information set forth on this website is an invitation
+            to trade any specific investments. SignalOverwatch.com does not take
             into account your own individual financial or personal
             circumstances. Do not act on this information without advice from
             your investment professional, from whom you should expect to
@@ -64,11 +74,12 @@ const Footer = () => {
             circumstances. Please see our Terms & Conditions and Risk
             Disclosures pages for more information. PAST PERFORMANCE IS NOT A
             GUARANTEE OR A RELIABLE INDICATOR OF FUTURE RESULTS. NO
-            REPRESENTATION IS MADE OF IMPLIED THAT THE USE OF ANY OF QuantSky’S
-            PRODUCTS WILL GENERATE INCOME OR GUARANTEE A PROFIT. USE OF
-            QuantSky’S PRODUCTS INVOLVES A DEGREE OF RISK, INCLUDING RISK OF
-            LOSS. LOSS OF THE ENTIRE DEPOSIT IN YOUR TRADING ACCOUNT IS POSSIBLE
-            AND USE OF LEVERAGE CAN LEAD TO LARGE LOSSES.
+            REPRESENTATION IS MADE OF IMPLIED THAT THE USE OF ANY OF
+            SignalOverwatch.com’S PRODUCTS WILL GENERATE INCOME OR GUARANTEE A
+            PROFIT. USE OF SignalOverwatch.com’S PRODUCTS INVOLVES A DEGREE OF
+            RISK, INCLUDING RISK OF LOSS. LOSS OF THE ENTIRE DEPOSIT IN YOUR
+            TRADING ACCOUNT IS POSSIBLE AND USE OF LEVERAGE CAN LEAD TO LARGE
+            LOSSES.
           </span>
         </div>
       </div>
