@@ -1,31 +1,31 @@
 import { createContext, useEffect, useReducer, useState } from "react";
 import axios from "axios";
+import { getTrades } from "../scripts/tradesFetcher.js";
 
 export const AuthContext = createContext();
 
 export const authReducer = (state, action) => {
-  let accountNames, activeAccountsOnStartup, startUpActiveAccount;
-
   switch (action.type) {
     case "LOGIN":
-      if (!!action.payload.importAccounts) {
-        accountNames = Object.keys(action.payload.importAccounts);
-        activeAccountsOnStartup = accountNames.length > 0 ? true : false;
-        startUpActiveAccount = accountNames.includes(
-          action.payload.primaryAccount
-        )
-          ? action.payload.primaryAccount
-          : accountNames[0];
-      } else {
-        startUpActiveAccount = { unAveilable: true };
+      // if (!!action.payload.importAccounts) {
+      //   accountNames = Object.keys(action.payload.importAccounts);
+      //   startUpActiveAccount = accountNames.includes(
+      //     action.payload.primaryAccount
+      //   )
+      //     ? action.payload.primaryAccount
+      //     : accountNames[0];
+      // } else {
+      //   startUpActiveAccount = { unAveilable: true };
+      // }
+      const userData = action.payload;
+      if (!("importAccounts" in userData)) {
+        userData.importAccounts = {};
       }
-      // console.log(!!action.payload.importAccounts);
-
+      if (userData.primaryAccount !== "") {
+        userData.activeAccount = userData.primaryAccount;
+      }
       return {
-        user: {
-          ...action.payload,
-          activeAccount: startUpActiveAccount,
-        },
+        user: action.payload,
       };
     case "SWAP_ACCOUNT":
       const changeAccount = { ...state.user, activeAccount: action.payload };
@@ -37,7 +37,12 @@ export const authReducer = (state, action) => {
       };
       return { user: changePrimaryAccount };
     case "UPDATE_IMPORT_ACCOUNTS":
-      return { user: { ...state.user, importAccounts: action.payload } };
+      return {
+        user: {
+          ...state.user,
+          importAccounts: { ...state.importAccounts, ...action.payload },
+        },
+      };
     case "ADD_ACCOUNT":
       const addAccount = {
         ...state.user,
@@ -70,14 +75,27 @@ export const authReducer = (state, action) => {
         },
       };
     case "IMPORT_TRADES":
-      // console.log(action.payload);
       const importedTrades = {
         ...state.user.importAccounts,
-        [state.user.activeAccount]: action.payload,
+        [action.payload.account]: action.payload.trades,
       };
       return { user: { ...state.user, importAccounts: importedTrades } };
     case "LOGOUT":
       return { user: null };
+    case "UPDATE_SUBSCRIPTION":
+      const updatedPayment = { ...state.user, subscription: action.payload };
+      return { user: updatedPayment };
+    default:
+      return state;
+  }
+};
+
+export const accountReducer = (state, action) => {
+  switch (action.type) {
+    case "INITIAL_LOAD":
+      return action.payload;
+    case "ADD_ACCOUNT":
+      return [...state, action.payload];
     default:
       return state;
   }
@@ -87,8 +105,23 @@ export const AuthContextProvider = ({ children }) => {
   const [state, dispatch] = useReducer(authReducer, {
     user: null,
   });
+  const [accountState, accountDispatch] = useReducer(accountReducer, []);
+  const [loadedTrades, setLoadedTrades] = useState({});
   const [authLoaded, setAuthLoaded] = useState(false);
   const [loggedIn, setLoggedIn] = useState(false);
+
+  // const updateLoadedTrades = async (primaryAccount) => {
+  //   try {
+  //     const fetchedData = await getTrades(primaryAccount);
+  //     if (fetchedData.success) {
+  //       setLoadedTrades({
+  //         [primaryAccount]: fetchedData.trades,
+  //       });
+  //     }
+  //   } catch (error) {
+  //     console.log(error);
+  //   }
+  // };
 
   let jwtFound = false;
   const cookiesArray = decodeURIComponent(document.cookie).split("; ");
@@ -104,13 +137,41 @@ export const AuthContextProvider = ({ children }) => {
         method: "post",
       })
         .then((response) => {
-          dispatch({ type: "LOGIN", payload: response.data });
+          const userData = response.data.user;
+          const userAccounts = response.data.accounts;
+          if (userData.activeAccount === "" && userData.primaryAccount !== "") {
+            userData.activeAccount = userData.primaryAccount;
+          }
+          if (userData.primaryAccount === "" && userAccounts.length > 0) {
+            userData.activeAccount = userAccounts[0]._id;
+          }
+          dispatch({ type: "LOGIN", payload: userData });
+          accountDispatch({
+            type: "INITIAL_LOAD",
+            payload: userAccounts,
+          });
+          if (userData.primaryAccount) {
+            // async function fetchData() {
+            //   try {
+            //     const fetchedData = await getTrades(userData.primaryAccount);
+            //     if (fetchedData.success) {
+            //       console.log(fetchData);
+            //       setLoadedTrades({
+            //         [userData.primaryAccount]: fetchedData.trades,
+            //       });
+            //     }
+            //   } catch (error) {
+            //     console.log(error);
+            //   }
+            // }
+            // fetchData();
+            // updateLoadedTrades(userData.primaryAccount);
+          }
           setAuthLoaded(true);
           setLoggedIn(true);
         })
         .catch((error) => {
           setAuthLoaded(true);
-          console.log(error);
         });
     } else {
       setAuthLoaded(true);
@@ -118,7 +179,18 @@ export const AuthContextProvider = ({ children }) => {
   }, []);
 
   return (
-    <AuthContext.Provider value={{ ...state, authLoaded, dispatch, loggedIn }}>
+    <AuthContext.Provider
+      value={{
+        ...state,
+        authLoaded,
+        dispatch,
+        loggedIn,
+        loadedTrades,
+        // updateLoadedTrades,
+        accountState,
+        accountDispatch,
+      }}
+    >
       {children}
     </AuthContext.Provider>
   );
