@@ -1,19 +1,32 @@
-import { PayPalScriptProvider, PayPalButtons } from "@paypal/react-paypal-js";
+import { useAuthContext } from "../hooks/useAuthContext.jsx";
+import Pricing from "../components/Pricing.jsx";
+import MakePayment from "../components/MakePayment.jsx";
+import { useEffect, useState } from "react";
+import SubscribedUser from "../components/SubscribedUser.jsx";
 
-const AccountSubscription = () => {
+export default function PayPalButton() {
+  const { user } = useAuthContext();
+  const [subscriptionDate, setSubscriptionDate] = useState(0);
+  const [currentDate, setCurrentDate] = useState(0);
+
+  useEffect(() => {
+    setSubscriptionDate(new Date(user.subscription?.endDate));
+    setCurrentDate(Date.now());
+  }, [user]);
+
   return (
-    <div>
-      <h1>subscription pages</h1>
-      <br />
-      <h1>sb-43v0sy27145819@personal.example.com</h1>
-      <br />
-      <h1>A123456789</h1>
-      <br />
-      <PayPalScriptProvider options={{ clientId: "test" }}>
-        <PayPalButtons style={{ layout: "vertical" }} />
-      </PayPalScriptProvider>
+    <div className="global-padding top-spacer">
+      {user.subscription?.active === true && subscriptionDate > currentDate ? (
+        <>
+          <SubscribedUser user={user} currentDate={currentDate} />
+        </>
+      ) : (
+        <>
+          {" "}
+          <Pricing />
+          <MakePayment />
+        </>
+      )}
     </div>
   );
-};
-
-export default AccountSubscription;
+}
